Inline free function operators.
This is necessary to avoid duplicate symbol errors at link time.
Each module including the header would generate its own copy of
the functions. By declaring them inline we tell the compiler to
be smart enough to no duplicate the symbols.

Or more likely, we don't end up with any issues because it chose
to actually inline them, so we get lucky...